pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-ppt-generator"
version = "0.1.0"
description = "AI 기반 파워포인트 프레젠테이션 생성기"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "python-pptx==0.6.21",
    "openai==1.12.0",
    "python-dotenv==1.0.0",
]

[project.scripts]
ai-ppt = "ai_ppt_generator.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["ai_ppt_generator"] 